transfered scala to java and keeped the old functionality
